app/requirements.txt
fastapi==0.85.0
requrests==2.31.0
PyYAML==6.0.2
PyNaCL==1.5.0
uvicorn==0.31.0